Add files via upload
On some computers, loading the userform was causing a crash: "Microsoft has stopped working. Windows can try recovering your information -> Close the program". The following actions were taken:

Separated Userform Initialization items into separate subs. ComboBox populations for Year, Grade and Division are now separare and called from within the initialization sub after the form is loaded. Did not fix the crash problem but probably better to have them separate anyway.

Tightened some of the loops for the combobox populating routines to a smaller range to speed things up a bit.

Determined that, for some strange reason, the crash would not occur if the VBA Editor was open. Added a line of code to open the VBA Editor when the file is opened. This seems to have fixed the problem.